refactor diagnostic formatter
